refactor tests constants
