--- src/Constants/testConstants.js
@@ -1,19 +1,0 @@
-const testConstants = {
-    validEmail: "maxime_labbe@outlook.com",
-    validFirstName: "Alfred",
-    validName: "Montmagny",
-    validPhoneNumber: "418-666-6666",
-    validPassword: "Jf3$E12",
-    invalidEmail: "a@a",
-    invalidPhoneNumber: "41-92-006",
-    invalidPassword: "44ab",
-    validNewPassword: "U51hy&6"
-};
-
-export default testConstants;
-
-export const testPopupConstant = {
-    validRGBColor: "#6aa84f",
-};
-
-
