+++ src/__tests__/testAddEmployee.test.js
@@ -5,7 +5,7 @@
 import "@testing-library/jest-dom";
 import {fireEvent, render} from "@testing-library/react";
 import {FormErrorType} from "../src/engine/messages/FormMessages";
-import testConstants from "../Constants/testConstants";
+import {testConstants} from "../Constants/testConstants";
 import userEvent from "@testing-library/user-event";
 import {MemoryRouter} from "react-router-dom";
 import {ComponentAddEmployee} from "../src/engine/components/ComponentAddEmployee";
