+++ src/__tests__/testComponentPopupSchedule.test.js
@@ -4,15 +4,10 @@
 
 import "@testing-library/jest-dom";
 import MatchMediaMock from "jest-matchmedia-mock";
-import { fireEvent, render } from "@testing-library/react";
-import { FormErrorType } from "../src/engine/messages/FormMessages";
+import { render } from "@testing-library/react";
 import testConstants from "../Constants/testConstants";
 import userEvent from "@testing-library/user-event";
 import { MemoryRouter } from "react-router-dom";
-import { act } from "react-dom/test-utils";
-import ReactDOM from 'react-dom/client';
-
-
 
 let user;
 let matchMedia;
